hardcoding the count of DKIM records is necessary.
Making it dependent on the list of tokens produces this error: The "count" value depends on resource attributes that cannot be determined until apply, so Terraform cannot predict how many instances will be created. To work around this, use the -target argument to first apply only the resources that the count depends on.